Add edge snapping, single-Esc return, and testable layout geometry
Snapping: panel edges snap to each other, to the wall edges, and to the wall
centre lines, with guide lines while a drag is live and Alt to defeat it.

It snaps twice on purpose. The overlay snaps in window pixels so the drag
tracks the hand, then the main process re-snaps the driven edge in wall units
before saving. Without the second pass an edge that looked snapped while
previewing a 4K layout on a laptop could save as 1919 against a neighbour's
1920, which is a visible seam at wall resolution. Verified: view-3 right edge
and view-4 left edge both land on exactly 1498.

Also stops cross-axis drift. A side drag round-tripped its untouched dimension
through window pixels and came back 1080 -> 1079, which would accumulate every
time the layout was touched. Only the dimensions a gesture actually drives are
now taken from the pointer; the rest are pinned to the drag baseline, and a
corner drag derives height from the baseline aspect exactly.

Esc: escToGrid is now "single" per Jeff. A single Esc returns to the grid, which
is what SPEC.md asks for and what anyone at the wall will expect. The tradeoff,
that dashboards using Esc for their own modals will not see the key, is recorded
in docs/validation.md along with the one-word revert. The policy also moved into
a single handleEscape() shared by the per-view handler and the overlay, which
had been deciding for itself and would have ignored the setting.

New src/layout.js holds the clamp and snap geometry as pure functions with no
electron import, covered by 17 tests in test/layout.test.js including one
asserting snapped neighbours share an exact edge. 34 tests total.

+++ src/config.js
@@ -125,9 +125,12 @@ function withDefaults(c) {
     hideInactiveWhenActive: !!c.hideInactiveWhenActive,
     transitionMs: c.transitionMs ?? 0,
     // Esc is ambiguous: the wall wants it for "back to grid", but real
-    // dashboards use it to close modals. "double" lets a single Esc reach the
-    // page and docks on a quick second press. See docs/validation.md.
-    escToGrid: c.escToGrid || 'double',
+    // dashboards also use it to close modals, and one key cannot do both.
+    // Jeff chose "single" (press Esc, return to grid) on 2026-08-21. If a real
+    // dashboard turns out to need Esc for its own modals, "double" lets the
+    // first press reach the page and docks on a quick second one. "off" leaves
+    // only the Back button and the idle timeout. See docs/validation.md.
+    escToGrid: c.escToGrid || 'single',
     escDoubleMs: c.escDoubleMs ?? 600,
     backButton: c.backButton || { x: 24, y: 24, width: 176, height: 56 },
     views: c.views.map((v, i) => ({
